save
working version, initiated formulas and conditional Data Records separation
