+++ uds/translator/data_record_definitions/__init__.py
@@ -23,8 +23,9 @@ from .did import (
     get_dids_2013,
     get_dids_2020,
 )
+from .conditional import CONDITIONAL_EVENT_TYPE_RECORD_09_2020,    EVENT_TYPE_RECORD_01, EVENT_TYPE_RECORD_09_2020
 from .dtc import (
-    CONDITIONAL_EVENT_TYPE_RECORD_09,
+
     DTC,
     DTC_AND_STATUS,
     DTC_COUNT,
@@ -47,8 +48,7 @@ from .dtc import (
     DTC_STORED_DATA_RECORDS_LIST_2013,
     DTC_STORED_DATA_RECORDS_LIST_2020,
     DTCS_WITH_STATUSES_AND_EXTENDED_DATA_RECORDS_DATA_LIST,
-    EVENT_TYPE_RECORD_01,
-    EVENT_TYPE_RECORD_09_2020,
+
     FAULT_DETECTION_COUNTER,
     GROUP_OF_DTC,
     MULTIPLE_DTC_AND_FAULT_DETECTION_COUNTERS,
